add probes for RF and emission WLG (#15)
Add probes that output the radiation field wavelength grid or the dust emission wavelength grid. The details of these wavelength grids are often useful when studying convergence (i.e. minimizing the number of wavelength bins) for a given input model. Even if these wavelength grids can be obtained as part of other probes, the extra info is in this case not needed (and it takes time to generate).