Build Voronoi grids using multiple parallel execution threads (#27)
The cells of Voronoi grids are now constructed using multiple parallel execution threads when the simulation is being run with a single process (see below). Also, the portions of the Voro++ code that are not used in SKIRT have been removed, and the formatting of the remaining Voro++ code has been streamlined.

When SKIRT is being run under MPI with multiple processes, building Voronoi grids reverts back to the old serial behaviour, and the percentage in the progress log is incorrect. This is because of an implementation limitation that would be tricky to remove. The plan is to fully support hydrid parallelisation in the near future, automatically resolving this problem.