Add facilities for defining spatial magnetic field distributions in the input model (#29)
We add facilities for defining spatial magnetic field distributions in the input model using some basic built-in vector fields and/or data imported from snapshots. The magnetic field will be used to determine the alignment direction of non-spherical grains in a future implementation.

The code runs the existing functional tests (and several new ones), so all changes are upwards compatible.