Avoid NaNs in (unused) precalculated data (#39)
Description
Fix a bug that caused NaN values for the mean mass of a dust population's size bin if the size bin does not contain any dust. The code depending on the mean mass happened to behave as if the NaN value was actually zero, so that the end result was correct anyway.

Motivation
Intermediate NaN values are evil. A future update to the code using the mean mass might have unsurfaced a very hard to find bug.

Context
Thanks to @drvdputt for reporting this (internal) bug.