minRadius feature for Sphere1DSpatialGrid (#41)
Description
Sphere1DSpatialGrid gets an extra property 'minRadius', analogous to 'maxRadius' of the parent class SphereSpatialGrid.

Motivation
This is useful for running thin spherical shell models, for which the inner cavity is relatively large. It avoid having to introduce a large number of empty shells, and makes it more intuitive to match the grid boundary with the medium boundary.

Context
This has been implemented in Sphere1DSpatialGrid, and not the parent class
SphereSpatialGrid, because the other subclass, Sphere2DSpatialGrid, is much more complicated and less frequently used.