Allow placing models at nonzero redshift; support dust heating by CMB (#43)
With this commit we support cosmology redshift in SKIRT. The simulated model can be placed at a nonzero redshift in a spatially flat cosmology with configurable parameters. Instruments can be placed in the model's restframe or in the observer frame, honouring the redshift. Dust heating by the cosmic microwave background (CMB) radiation at the model's redshift can be added to the regular radiation field.

Although the new features seem to produce reasonable results, please don't use them for 'production' simulations until we have had a chance to verify their accuracy in more detail.