Simplify mechanism for allocating random number generators
Description
Simplify mechanism for allocating random number generators (RNGs) to threads. The main thread gets a predictable RNG using the user configured seed. Child threads get a RNG with an arbitrary, truly random seed. When executed in single thread/single process mode, everything is performed in the main thread, so that the random sequence is predictable. Otherwise, serial tasks are performed in the main thread, and all parallelized tasks are performed in a child thread (because the main thread is not used to do actual work it is not counted against the number of threads specified on the command line). This is also true if each of the multiple processes uses only a single worker threads.

Motivation
There was suboptimal randomization when performing SKIRT with just a single thread in each of multiple processes (a less frequently used mode anyway). Furthermore, removing some sections of complex source code and comments is always a good thing...