Fix bug in optimization of instruments with the same line of sight (#59)
Description
This change fixes a bug in the optimization of consecutive instruments with the same line of sight. In some cases, an instrument would incorrectly use the line of sight of the previous instrument instead of its own line of sight.

Motivation
The previous behavior was incorrect.

Tests
All functional tests still run and a new test has been added for this particular problem.

Context
Thank you to @anatrcka for reporting this problem.