Warning message for incorrect installed resource pack version (#60)
SKIRT now logs a warning message both on the console and in each simulation's log file when the Core resource pack is not installed or the version number of any installed resource pack does not match the expected version.

Motivation: we will shortly be releasing a new version of the Core resource pack with a small change that is hard to detect without verifying the version number. It is therefore important to notify users if they have the incorrect version installed.