Lyman-alpha line transfer capability (#64)
This pull request adds Lyman-alpha line transfer capability to SKIRT, including the photon life cycle and the ability to import a neutral hydrogen density distribution from snapshots. This is a progression of the Lyman-alpha implementation in SKIRT that was started a while ago by adding some Lyman-alpha specific SEDs and SED families.

Important caveat
The overhead caused by the forced scattering mechanism used in SKIRT seems to be make the runtime for realistic Lyα models prohibitively long. We will need to implement an alternative photon cycle that does not employ forced scattering. This is tricky because currently the calculations for the path geometry and for the optical depth happen in two separate phases that cannot be easily intertwined.