HEALPix sky instrument (#65)
This commit implements a new Instrument, called HEALPixSkyInstrument. This Instrument is similar to the existing AllSkyInstrument, but instead of using a projection to map its output internally to a flat image, it directly outputs the pixel information onto a uniform pixelisation of the celestial sphere without any internal projection.

The existing AllSkyInstrument does not sample the sky uniformly, which makes it hard to compare pixels in different parts of the sky during analysis. It is also very hard to change the orientation of the resulting image in post-processing. A HEALPix pixelisation of the sky has the advantage that each pixel samples exactly the same fraction of the sky, that distances between HEALPix pixels are more uniform, and that it is possible to rotate the sky image in post-processing to e.g. put one of the polar regions in the centre of the image. This kind of pixelisation is also more natural when comparing SKIRT images with the full-sky images of the Planck satellite.

See pull request #65