Correctly sample the anisotropic polarised emission profile (#71)
Made sure anisotropic polarised secondary emission is sampled correctly and is transformed to the correct reference frame. There were two issues: first of all, there was a small issue with the calculation of the cumulative distribution function for the sampling. Second and most important, in the old implementation the randomly generated direction was not transformed from the dust frame (reference frame with the z-axis along the magnetic field direction) to the lab frame, leading to a completely incorrect sampling of the anisotropic emission profile.