Add snapshot import from list of cuboidal cells (as simpler alternative for AMR) (#73)
This pull request adds the the capability to import snapshot data (for media and sources) from a list of cuboidal cells. The intention is for the cells to define a partition of the spatial domain, however this is not enforced. This new functionality is an alternative for using the adaptive mesh import capability, which requires listing cells in Morton order. This can be hard to accomplish, especially when extracting a particular subdomain from a larger mesh. In contrast, the new option allows cells to be listed in arbitrary order and allows the spatial partition to be imperfect. For example, the outer border of the domain can be ragged, with some cells extending farther out than others.