Added a 3D geometry read from a .fits datacube
Added the functionality to read the spatial luminosity distribution for sources or the spatial density distribution for the medium from a .fits datacube. This allows SKIRT users to introduce arbitrary, asymmetric geometries into the simulation. The model geometry is set by two parameters: the input filename and the pixelscale (i.e. the physical length per pixel).