Fix the asymmetry in Lyman-alpha SEDs for static spheres (#81)
This commit fixes the slight but systematic asymmetry noted in Lyman-alpha SEDs for static sphere benchmarks. The culprit was a bug in the implementation of the choice between isotropic and dipole scattering phase functions (the dipole was selected only for wavelengths on one side of the Lyman-alpha line center).

Also, a minor performance optimization was made in the calculation of the wavelength shift for each scattering event.