Avoid construction of Voronoi tessellation for some use cases (#95)
This update offers the possibility to forego the construction of the Voronoi tessellation for imported VoronoiMesh media in case the full tessellation is not needed for purposes other than sampling the density distribution. An important use case is when the medium density distribution defined on the Voronoi grid is resampled to an octree grid to perform radiative transfer. To enable this use case, the massType option can be set to include both mass density and volume-integrated mass (or both number density and volume-integrated number) in the imported data file.