Add probe for radiation field at specified positions (#96)
Add a probe that outputs the mean radiation field intensity at the positions listed in a user-provided input file. This allows, for example, to probe the radiation field at the positions listed in a Particle or VoronoiMesh import file used for defining sources or media by configuring this new probe with the same import file (any columns other than the position coordinates will be ignored).