add an option for considering the doppler shift of dust emission.
I faced a problem: Dust continuum around the emission lines from the gas in the spectrum becomes smaller, when I perform non-LTE line + dust radiative transfer . This is because the Doppler shift causes a difference between the wavelength range for each line of the instrument and that of the dust emission. Furthermore, MediumState cannot consider different velocity fields of each mediums. That means that if we want to consider gas emission line, we always have to consider the doppler shift of dust emission. To avoid this problem, an option for doppler shift of dust emission is included.